add doc. move fp6_llm under prototype
